# pkg/docs/swagger.yaml
definitions:
  ginx.PaginatedResp:
    properties:
      count:
        type: integer
      results: {}
    type: object
  ginx.Response:
    properties:
      data: {}
      message:
        type: string
      requestID:
        type: string
    type: object
  multipart.FileHeader:
    properties:
      filename:
        type: string
      header:
        $ref: '#/definitions/textproto.MIMEHeader'
      size:
        type: integer
    type: object
...
  serializer.CategoryListResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
      updater:
        type: string
    type: object
...
  serializer.CreateDirRequest:
    properties:
      dirPath:
        type: string
    required:
    - dirPath
    type: object
...
  serializer.UploadObjectRequest:
    properties:
      dirPath:
        type: string
      file:
        $ref: '#/definitions/multipart.FileHeader'
    type: object
...
  textproto.MIMEHeader:
    additionalProperties:
      items:
        type: string
      type: array
    type: object
info:
  contact: {}
  description: 蓝鲸开发者中心 SaaS 后台 API 文档
  title: BkApp API DOC
  version: 1.0.0
paths:
...
  /api/obj-storage/dirs:
    delete:
      parameters:
      - in: query
        name: dirPath
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: 删除目录
      tags:
      - object-storage
    post:
      parameters:
      - description: 创建目录请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/serializer.CreateDirRequest'
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Response'
            - properties:
                data:
                  type: object
              type: object
      summary: 创建目录
      tags:
      - object-storage
  /api/obj-storage/objects:
    delete:
      parameters:
      - in: query
        name: dirPath
        type: string
      - in: query
        name: objName
        type: string
      responses:
        "204":
          description: No Content
      summary: 删除已上传对象
      tags:
      - object-storage
    get:
      parameters:
      - in: query
        name: dirPath
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/ginx.PaginatedResp'
                  - properties:
                      results:
                        items:
                          $ref: '#/definitions/serializer.CategoryListResponse'
                        type: array
                    type: object
              type: object
      summary: 获取已上传对象列表
      tags:
      - object-storage
    post:
      parameters:
      - description: 上传对象请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/serializer.UploadObjectRequest'
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Response'
            - properties:
                data:
                  type: object
              type: object
      summary: 上传对象
      tags:
      - object-storage
  /api/obj-storage/objects/download:
    get:
      parameters:
      - in: query
        name: dirPath
        type: string
      - in: query
        name: objName
        type: string
      responses:
        "200":
          description: 文件内容
          schema:
            type: file
      summary: 下载对象
      tags:
      - object-storage
...
swagger: "2.0"